requirements.txt
fastapi
uvicorn
pydantic
pika
python-dotenv
python-multipart
sqlalchemy
pymysql
